complete telegram auth + refactoring entities

+++ backend/security/src/main/kotlin/ru/sigma/security/service/JwtService.kt
@@ -74,7 +74,7 @@ class JwtService(
 
         return TokenInfo(
             token = token,
-            expiresIn = liveTime.toSeconds().toInt()
+            expiresIn = liveTime.toSeconds()
         )
     }
